Rename Nockma stdlib to anomalib and add RM references (#3111)
Most changes in this PR relate to renaming of the Anoma Nock
StandardLibrary to AnomaLibrary. This is because the Anoma library now
consists of a standard library from
[anoma.hoon](https://github.com/anoma/anoma/blob/a20b5e78380b4dcf554fda6f4c75758cc1b70063/hoon/anoma.hoon)
and the resource machine library
[resource-machine.hoon](https://github.com/anoma/anoma/blob/a20b5e78380b4dcf554fda6f4c75758cc1b70063/hoon/anoma.hoon).

The Anoma RM functions and value references are also added. Their
integration into the compiler pipeline will happen in a separate PR.

The Anoma Library RM functions and stdlib functions are enumerated
separately. There is a separate type for Anoma Library values because
these are compiled differently than functions.

Part of:

* https://github.com/anoma/juvix/issues/3084