Launch the Anoma node and client from the elixir REPL
libsodium cryptographic APIs like sign-detached cannot currently be
called from within the Anoma node or client binaries.

Until this is solved we must start both the node and client from the
elixir REPL.

Previously we were starting the node using the REPL and the client using
the binary.

This commit changes the `start.exs` script we were using to start the
node to now start both a node and a client.