Add a new Anoma compile test mode 'AnomaTestModeNodeOnly'
This mode is for tests that must run on the Anoma Node / Client. These
tests contain calls that are crashes in AnomaLib nock file and must
therefore be jetted.

With this changed the contents of test085 (resource machine builtins)
can be uncommnented. However it currently fails, unexpectedly on the node.