Fix source locations for flatparse backend and add tests for import scanning (#3265)
It is a current issues in Juvix that when an imports points to a missing
file, the error is reported in an incorrect location. This pr fixes that
and adds tests for it.
Bug example:

![image](https://github.com/user-attachments/assets/8c8ca527-4d64-47a1-80f3-c9de2535dd6d)
After the fix:

![image](https://github.com/user-attachments/assets/a7adf759-0581-4c88-a1c8-5f8f6e1dfa3f)


Changes included in this pr:
1. Fixes issues related to source locations. The flatparse line and
column numbers are 0-based, whereas megaparsec uses 1-based for line
numbers and 0 based for column numbers. Moreover, flatparse uses byte
count offset whereas megaparsec uses character count offset.
2. All scoping tests now have an additional check. We check that the
imports scanned using megaparsec and flatparse coincide, including the
source location.
3. I've added a flag `--check` for the `dev import-tree scan` command
that checks that all backends agree on the scanned imports in the given
file. If they don't, it outputs diff-friendly yaml files.