Fix let-patterns inside named arguments (#3274)
* Closes #3269 
* The problem was that `NameKind` of a name used as a variable was not
properly updated to `KNameLocal`, which later caused the identifier to
be converted into `IdenFunction` by `substitutionE` called by the code
handling named arguments. The function `substitutionE` calls
`toExpression` on the name instead of on the identifier expression, and
`toExpression` called on a name chooses the kind of identifier to create
based on `NameKind`. In my opinion, it is error-prone to have two
distinct places which determine the kind of an identifier. In
`genPatternDefs`, I changed `IdenFunction` to `IdenVar` but forgot about
adjusting `NameKind`.