Remove mainFunctionWrapper in nock backend
THe mainFunctionWrapper was used to capture the argument of main to use
as the `id` field of a anomaGet (scry) call.

The `id` argument is no longer passed into the main function in all
cases. anomaGet must support being called with the `id` argument
explicitly, it is now part of the scry key.