Avoid double jamming prove arguments in Anoma CLI commands (#3298)
This PR changes the way program arguments are processed by the `juvix
dev nockma run {with-client, ephemeral-client}` and `juvix dev anoma
prove` CLI commands.

Files provided in `--arg 'bytes:..'` and `--arg 'base64:..'` flags are
now assumed to be jammed and are not jammed again before they are
submitted to the Anoma client.

## Changes in anoma-apps

The Anoma apps in anoma-apps use the `prove` CLI command and so the
signatures of their `main` functions will change:

For example, in
[HelloWorld.juvix](https://github.com/anoma/anoma-apps/blob/6132630209dacb47b8511b2d5b60859af46f2437/HelloWorld/HelloWorld.juvix)
the logic is passed as a jammed noun in an argument to prove.

Previously the argument was jammed again before being sent and so had to
be decoded in the `main` function before being used (the Anoma client
cues (i.e decodes) the argument once before calling the program).

```
main (encodedLogic : Encoded Logic) (message : String) : TransactionRequest
```

After the change in this PR, the logic argument is not jammed again
before being sent to the Anoma client and so the signature of the main
function should be changed to:

```
main (logic : Logic) (message : String) : TransactionRequest
```