Fix bug in scoping for default arguments (#3318)
when we have:
```
f {arg : Nat := X} ...
```
then `arg` should not be in scope in `X`. 

Before this pr, if `arg` was used in `X`, we got an impossible error in
the typechecker.
E.g. typechecking the following results in a crash. After this pr, it
gives a scoping error.
```
module bug;

type T := mkT;

f {x : T := x} : T := mkT;
```