Modular compilation in Core (#3315)
Extends modularization to the Core pipelines. This is still based on the
existing modularization framework which goes down to Core. Now the
"stored Core" format depends on the target and chosen options, e.g.,
with optimizations enabled the code in each stored Core module will be
optimized. Different subdirectories within the build directory are used
for different formats (so, e.g., the entire project doesn't need to be
recompiled when switching from a release to a debug build, if there was
a debug build saved before). Also, typechecking is now a bit faster,
because the "stored Core" format for the typechecking pipeline includes
only the transformations necessary for typechecking (earlier all
`PipelineStored` transformations were run on the dependency modules).