Accept optional semicolon before pipe (#3334)
I often make this mistake:
```
case x of 
  | p := a; -- this semicolon gives an error
  | q := b;
```
Even if you always remember this, sometimes you have to manually delete
the semicolon if you reorder the case branches (.e.g. if the last
becomes second to last).

This pr changes the parser so that an optional semicolon before any `|`
is accepted. The semicolon is discarded so it will be removed by the
formatter