Check package name conflicts (#3352)
- Progress towards #3331

This pr adds a check that ensures no two packages in the dependency tree
have the same name but different versions. Example error message:
```
/home/jan/projects/juvix/tests/negative/PackageNameConflict/dep2/Package.juvix:1:1: error:
The package dep is used with different versions:
• version 0.11.0-5075d2e2d2ea0ceedb49a53ad8daf559c89cb9c3f86b11853bcc1586907cf3d1
  • at /home/jan/projects/juvix/tests/negative/PackageNameConflict/dep2/
• version 1.0.0-5075d2e2d2ea0ceedb49a53ad8daf559c89cb9c3f86b11853bcc1586907cf3d1
  • at /home/jan/projects/juvix/tests/negative/PackageNameConflict/dep1/
```

I've added a global flag that disables this check:
`--unsafe-ignore-package-name-conflicts`. I've added this flag just in
case some existing Juvix project relied on the existing behaviour. We
should probably remove it in the near future.