Fix modular compilation with transitive imports (#3396)
The number of scry calls to load imports used to be exponential, because
loading a module triggered recursive loading of its dependencies (count
the number of paths in the import DAG). For the standard library, it was
almost 2 million scry calls.

This PR changes the implementation of modularization to fetch a module
when a function from that module is called, instead of prefetching them
on startup.

Also changes the module format to make it more convenient for submission
to the Anoma client.